Docker Volume for yaml database weight config files
Docker Compose will load a folder called config/ if it is present in
the user's directory beside the compose.yaml file when the user runs
`docker compose up`.

how yaml files are loaded now
-----
Each XXX_db_entry.py handles it's own weights, so you don't need to add
any protein weight files to settings or python files.
It now works by the protein entries defining a default dictionary for
their weights that they expect and will use in their scoring
functions.
weight_importer.py will try to load a specified config file (in the
docker volume folder), if this doesn't exist it will use the default
weights.
If the yaml file does exist, then the yaml dict and the default dict
will be combined (where yaml keys are lowercased, and override default
values when they exist)
This dict is then returned and used by db_entry in the same manner as
the default dict would have been.

In this way the code doesnt care if config files exist or not, and the
scoring functions can guarentee which weights it will have access to
by writing whatever default weight dictionary it wants.

reorganisation of codebase
------
I moved the databse entry types to their own folder, removed the
scoring weights folder (as this functionailty is moved to the entries
themselves), and made the filenames more consitant.
That is, python files are snake and classes are still Pascal case.

so before we would have AFDBentry.AFDBentry
but now it would be afdb_entry.AFDBentry